+++ UnitTests/BraintreeApplePayTests/BTApplePay_Tests.swift
@@ -1,6 +1,7 @@
 import XCTest
 import PassKit
 @testable import BraintreeTestShared
+@testable import BraintreeApplePay
 
 class BTApplePay_Tests: XCTestCase {
     var mockClient : MockAPIClient = MockAPIClient(authorization: "development_tokenization_key")!
@@ -23,8 +24,9 @@ class BTApplePay_Tests: XCTestCase {
         let expectation = self.expectation(description: "Callback invoked")
         applePayClient.paymentRequest { (paymentRequest, error) in
             guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTApplePayErrorDomain)
-            XCTAssertEqual(error.code, BTApplePayErrorType.unsupported.rawValue)
+            XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
+            XCTAssertEqual(error.code, BTApplePayError.unsupported.rawValue)
+            XCTAssertEqual(error.localizedDescription, BTApplePayError.unsupported.errorDescription)
             expectation.fulfill()
         }
 
@@ -38,24 +40,9 @@ class BTApplePay_Tests: XCTestCase {
         let expectation = self.expectation(description: "Callback invoked")
         applePayClient.paymentRequest { (paymentRequest, error) in
             guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTApplePayErrorDomain)
-            XCTAssertEqual(error.code, BTApplePayErrorType.unsupported.rawValue)
-            expectation.fulfill()
-        }
-
-        waitForExpectations(timeout: 2, handler: nil)
-    }
-
-    func testPaymentRequest_whenAPIClientIsNil_callsBackWithError() {
-        let applePayClient = BTApplePayClient(apiClient: mockClient)
-        applePayClient.apiClient = nil
-
-        let expectation = self.expectation(description: "Callback invoked")
-        applePayClient.paymentRequest { (paymentRequest, error) in
-            XCTAssertNil(paymentRequest)
-            guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTApplePayErrorDomain)
-            XCTAssertEqual(error.code, BTApplePayErrorType.integration.rawValue)
+            XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
+            XCTAssertEqual(error.code, BTApplePayError.unsupported.rawValue)
+            XCTAssertEqual(error.localizedDescription, BTApplePayError.unsupported.errorDescription)
             expectation.fulfill()
         }
 
@@ -130,10 +117,11 @@ class BTApplePay_Tests: XCTestCase {
 
         let client = BTApplePayClient(apiClient: mockClient)
         let payment = MockPKPayment()
-        client.tokenizeApplePay(payment) { (tokenizedPayment, error) -> Void in
+        client.tokenize(payment) { (tokenizedPayment, error) -> Void in
             guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTApplePayErrorDomain)
-            XCTAssertEqual(error.code, BTApplePayErrorType.unsupported.rawValue)
+            XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
+            XCTAssertEqual(error.code, BTApplePayError.unsupported.rawValue)
+            XCTAssertEqual(error.localizedDescription, BTApplePayError.unsupported.errorDescription)
             expectation.fulfill()
         }
 
@@ -146,26 +134,11 @@ class BTApplePay_Tests: XCTestCase {
 
         let client = BTApplePayClient(apiClient: mockClient)
         let payment = MockPKPayment()
-        client.tokenizeApplePay(payment) { (tokenizedPayment, error) -> Void in
+        client.tokenize(payment) { (tokenizedPayment, error) -> Void in
             guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTApplePayErrorDomain)
-            XCTAssertEqual(error.code, BTApplePayErrorType.unsupported.rawValue)
-            expectation.fulfill()
-        }
-
-        waitForExpectations(timeout: 2, handler: nil)
-    }
-
-    func testTokenization_whenAPIClientIsNil_callsBackWithError() {
-        let client = BTApplePayClient(apiClient: mockClient)
-        client.apiClient = nil
-
-        let expectation = self.expectation(description: "Callback invoked")
-        client.tokenizeApplePay(MockPKPayment()) { (tokenizedPayment, error) -> Void in
-            XCTAssertNil(tokenizedPayment)
-            guard let error = error as NSError? else {return}
-            XCTAssertEqual(error.domain, BTApplePayErrorDomain)
-            XCTAssertEqual(error.code, BTApplePayErrorType.integration.rawValue)
+            XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
+            XCTAssertEqual(error.code, BTApplePayError.unsupported.rawValue)
+            XCTAssertEqual(error.localizedDescription, BTApplePayError.unsupported.errorDescription)
             expectation.fulfill()
         }
 
@@ -178,7 +151,7 @@ class BTApplePay_Tests: XCTestCase {
         let payment = MockPKPayment()
         let expectation = self.expectation(description: "tokenization error")
 
-        client.tokenizeApplePay(payment) { (tokenizedPayment, error) -> Void in
+        client.tokenize(payment) { (tokenizedPayment, error) -> Void in
             guard let error = error as NSError? else {return}
             XCTAssertEqual(error.domain, "MyError")
             XCTAssertEqual(error.code, 1)
@@ -200,7 +173,7 @@ class BTApplePay_Tests: XCTestCase {
         let payment = MockPKPayment()
         let expectation = self.expectation(description: "tokenization failure")
 
-        client.tokenizeApplePay(payment) { (tokenizedPayment, error) -> Void in
+        client.tokenize(payment) { (tokenizedPayment, error) -> Void in
             XCTAssertEqual(error! as NSError, self.mockClient.cannedResponseError!)
             expectation.fulfill()
         }
@@ -219,7 +192,7 @@ class BTApplePay_Tests: XCTestCase {
         let payment = MockPKPayment()
         let expectation = self.expectation(description: "tokenization failure")
 
-        client.tokenizeApplePay(payment) { (tokenizedPayment, error) -> Void in
+        client.tokenize(payment) { (tokenizedPayment, error) -> Void in
             guard let error = error as NSError? else {return}
             XCTAssertEqual(error.domain, "MyError")
             XCTAssertEqual(error.code, 1)
@@ -258,7 +231,7 @@ class BTApplePay_Tests: XCTestCase {
 
         let client = BTApplePayClient(apiClient: mockClient)
         let payment = MockPKPayment()
-        client.tokenizeApplePay(payment) { (tokenizedPayment, error) -> Void in
+        client.tokenize(payment) { (tokenizedPayment, error) -> Void in
             XCTAssertNil(error)
             XCTAssertEqual(tokenizedPayment!.nonce, "an-apple-pay-nonce")
             XCTAssertTrue(tokenizedPayment!.isDefault)
@@ -279,7 +252,7 @@ class BTApplePay_Tests: XCTestCase {
         waitForExpectations(timeout: 2, handler: nil)
     }
     
-    func testTokenizeApplePay_whenNetworkConnectionLost_sendsAnalytics() {
+    func testTokenize_whenNetworkConnectionLost_sendsAnalytics() {
         mockClient.cannedResponseError = NSError(domain: NSURLErrorDomain, code: -1005, userInfo: [NSLocalizedDescriptionKey: "The network connection was lost."])
         
         mockClient.cannedConfigurationResponseBody = BTJSON(value: [
@@ -292,13 +265,55 @@ class BTApplePay_Tests: XCTestCase {
         let payment = MockPKPayment()
 
         let expectation = self.expectation(description: "Callback invoked")
-        applePayClient.tokenizeApplePay(payment) { nonce, error in
+        applePayClient.tokenize(payment) { nonce, error in
             XCTAssertNotNil(error)
             expectation.fulfill()
         }
         waitForExpectations(timeout: 2)
         
         XCTAssertTrue(mockClient.postedAnalyticsEvents.contains("ios.apple-pay.network-connection.failure"))
+    }
+
+    func testTokenize_whenBodyIsMissingData_returnsError() {
+        mockClient.cannedConfigurationResponseBody = BTJSON(value: ["applePay" : ["status" : "production"]])
+        mockClient.cannedResponseBody = nil
+
+        let applePayClient = BTApplePayClient(apiClient: mockClient)
+        let payment = MockPKPayment()
+        let expectation = expectation(description: "Callback invoked")
+
+        applePayClient.tokenize(payment) { nonce, error in
+            XCTAssertNil(nonce)
+            XCTAssertNotNil(error)
+            guard let error = error as NSError? else { XCTFail("Should return error"); return }
+            XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
+            XCTAssertEqual(error.code, BTApplePayError.noApplePayCardsReturned.rawValue)
+            XCTAssertEqual(error.localizedDescription, BTApplePayError.noApplePayCardsReturned.errorDescription)
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 2)
+    }
+
+    func testTokenize_bodyDoesNotContainApplePayCards_returnsError() {
+        mockClient.cannedConfigurationResponseBody = BTJSON(value: ["applePay" : ["status" : "production"]])
+        mockClient.cannedResponseBody = BTJSON(value: ["notApplePayCards": "badData"])
+
+        let applePayClient = BTApplePayClient(apiClient: mockClient)
+        let payment = MockPKPayment()
+        let expectation = expectation(description: "Callback invoked")
+
+        applePayClient.tokenize(payment) { nonce, error in
+            XCTAssertNil(nonce)
+            XCTAssertNotNil(error)
+            guard let error = error as NSError? else { XCTFail("Should return error"); return }
+            XCTAssertEqual(error.domain, BTApplePayError.errorDomain)
+            XCTAssertEqual(error.code, BTApplePayError.failedToCreateNonce.rawValue)
+            XCTAssertEqual(error.localizedDescription, BTApplePayError.failedToCreateNonce.errorDescription)
+            expectation.fulfill()
+        }
+
+        waitForExpectations(timeout: 2)
     }
 
     // MARK: - Metadata
@@ -314,7 +329,7 @@ class BTApplePay_Tests: XCTestCase {
         let payment = MockPKPayment()
 
         let expectation = self.expectation(description: "Tokenized card")
-        applePayClient.tokenizeApplePay(payment) { _,_  -> Void in
+        applePayClient.tokenize(payment) { _,_  -> Void in
             expectation.fulfill()
         }