+++ UnitTests/BraintreeApplePayTests/BraintreeApplePayTests-Bridging-Header.h
@@ -1,3 +1,1 @@
-#import <BraintreeApplePay/BTApplePayClient_Internal.h>
-#import <BraintreeApplePay/BTApplePayCardNonce.h>
 #import <BraintreeCore/BraintreeCore-Swift.h>